+++ lib/aozora2html/version.rb
@@ -1,3 +1,3 @@
 class Aozora2Html
-  VERSION = "0.7.1"
+  VERSION = "0.9.0"
 end
